refactor: rename testutil module to test
