+++ .idea/encodings.xml
@@ -9,7 +9,7 @@
     <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/test-utils/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/test-utils/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/testutil/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/testutil/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/test/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/test/src/main/resources" charset="UTF-8" />
   </component>
 </project>
