+++ lib/src/test/java/ch/sbb/pfi/netzgrafikeditor/converter/core/NetworkGraphicConverterIT.java
@@ -11,10 +11,10 @@ import ch.sbb.pfi.netzgrafikeditor.converter.core.supply.fallback.NoVehicleCircu
 import ch.sbb.pfi.netzgrafikeditor.converter.io.gtfs.GtfsScheduleWriter;
 import ch.sbb.pfi.netzgrafikeditor.converter.io.matsim.TransitScheduleXmlWriter;
 import ch.sbb.pfi.netzgrafikeditor.converter.io.netzgrafik.JsonFileReader;
-import ch.sbb.pfi.netzgrafikeditor.converter.util.test.OutputDir;
-import ch.sbb.pfi.netzgrafikeditor.converter.util.test.TestCase;
-import ch.sbb.pfi.netzgrafikeditor.converter.util.test.TestDirectoryExtension;
-import ch.sbb.pfi.netzgrafikeditor.converter.util.test.TestScenario;
+import ch.sbb.pfi.netzgrafikeditor.converter.test.OutputDir;
+import ch.sbb.pfi.netzgrafikeditor.converter.test.TestCase;
+import ch.sbb.pfi.netzgrafikeditor.converter.test.TestDirectoryExtension;
+import ch.sbb.pfi.netzgrafikeditor.converter.test.TestScenario;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.extension.ExtendWith;
