+++ lib/src/test/java/ch/sbb/pfi/netzgrafikeditor/converter/io/csv/CsvInfrastructureRepositoryTest.java
@@ -2,7 +2,7 @@ package ch.sbb.pfi.netzgrafikeditor.converter.io.csv;
 
 import ch.sbb.pfi.netzgrafikeditor.converter.core.supply.StopFacilityInfo;
 import ch.sbb.pfi.netzgrafikeditor.converter.core.supply.TrackSegmentInfo;
-import ch.sbb.pfi.netzgrafikeditor.converter.util.test.TestScenario;
+import ch.sbb.pfi.netzgrafikeditor.converter.test.TestScenario;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
