+++ test/src/main/java/ch/sbb/pfi/netzgrafikeditor/converter/test/InputRootDir.java
@@ -1,4 +1,4 @@
-package ch.sbb.pfi.netzgrafikeditor.converter.util.test;
+package ch.sbb.pfi.netzgrafikeditor.converter.test;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
