+++ test/src/main/java/ch/sbb/pfi/netzgrafikeditor/converter/test/OutputDir.java
@@ -1,8 +1,8 @@
-package ch.sbb.pfi.netzgrafikeditor.converter.util.test;
+package ch.sbb.pfi.netzgrafikeditor.converter.test;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-public @interface InputDir {
+public @interface OutputDir {
 }
