+++ test/src/main/java/ch/sbb/pfi/netzgrafikeditor/converter/test/TestCase.java
@@ -1,4 +1,4 @@
-package ch.sbb.pfi.netzgrafikeditor.converter.util.test;
+package ch.sbb.pfi.netzgrafikeditor.converter.test;
 
 import lombok.Getter;
