+++ test/src/test/resources/ng/cases/cycle-small.json
