+++ test/src/test/resources/ng/cases/cycle.json
