+++ test/src/test/resources/ng/cases/self-intersection-random-order.json
