+++ test/src/test/resources/ng/cases/self-intersection.json
