+++ test/src/test/resources/ng/cases/self-overlay-random-order.json
