+++ test/src/test/resources/ng/cases/self-overlay.json
