+++ test/src/test/resources/ng/cases/simple-double-pass.json
