+++ test/src/test/resources/ng/cases/simple-dwell-time-conflict.json
