+++ test/src/test/resources/ng/cases/simple-pass.json
