+++ test/src/test/resources/ng/cases/simple.json
